requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
datasets>=2.14.0
numpy>=1.24.0
torch>=2.0.0
matplotlib>=3.7.0
psutil>=5.9.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
jupyter>=1.0.0
seaborn>=0.12.0
